cv_analysis_backend/apis-main/requirements.txt
Flask==2.3.2
pandas==1.5.3
python-dotenv==1.0.0
PyMuPDF==1.21.0
python-docx==0.8.11
psycopg2-binary==2.9.6
Werkzeug==2.3.3
scikit-learn==1.4.0
flask-cors==4.0.1
PyJWT
mongoengine
pymongo
Flask_login